+++ braindecode/models/labram.py
@@ -822,8 +822,9 @@ class Labram(EEGModuleMixin, nn.Module):
         )
 
     def reset_head(self, n_outputs):
-        self._n_outputs = n_outputs
+        self._set_n_outputs(n_outputs)
         self.reset_classifier(n_outputs)
+        self.final_layer.train(self.training)
 
     def _adj_temporal_embedding(self, num_ch, batch_size, dim_embed=None):
         """