+++ braindecode/models/signal_jepa.py
@@ -642,7 +642,7 @@ class SignalJEPA_Contextual(_BaseSignalJEPA):
         )
 
     def reset_head(self, n_outputs):
-        self._n_outputs = n_outputs
+        self._set_n_outputs(n_outputs)
         self.final_layer = _get_separable_clf_layer(
             conv_layers_spec=self._clf_conv_layers_spec,
             n_chans=self.n_chans,
@@ -650,6 +650,7 @@ class SignalJEPA_Contextual(_BaseSignalJEPA):
             n_classes=n_outputs,
             n_spat_filters=self._clf_n_spat_filters,
         )
+        self.final_layer.train(self.training)
 
     def forward(self, X, return_features=False):  # type: ignore
         local_features = self.feature_encoder(X)  # type: ignore
@@ -886,7 +887,7 @@ class SignalJEPA_PostLocal(_BaseSignalJEPA):
         )
 
     def reset_head(self, n_outputs):
-        self._n_outputs = n_outputs
+        self._set_n_outputs(n_outputs)
         self.final_layer = _get_separable_clf_layer(
             conv_layers_spec=self._clf_conv_layers_spec,
             n_chans=self.n_chans,
@@ -894,6 +895,7 @@ class SignalJEPA_PostLocal(_BaseSignalJEPA):
             n_classes=n_outputs,
             n_spat_filters=self._clf_n_spat_filters,
         )
+        self.final_layer.train(self.training)
 
     @classmethod
     def from_pretrained(
@@ -1119,11 +1121,12 @@ class SignalJEPA_PreLocal(_BaseSignalJEPA):
         )
 
     def reset_head(self, n_outputs):
-        self._n_outputs = n_outputs
+        self._set_n_outputs(n_outputs)
         self.final_layer = nn.Sequential(
             nn.Flatten(start_dim=1),
             nn.Linear(self._out_emb_dim, n_outputs),
         )
+        self.final_layer.train(self.training)
 
     @classmethod
     def from_pretrained(